posttest 7

+++ delete.php
@@ -1,6 +1,15 @@
 <?php require("includes/connection.php") ?>
 
 <?php
+session_start();
+
+if (!isset($_SESSION['login'])) {
+    header("Location: login.php");
+    exit;
+}
+?>
+
+<?php
 if (isset($_GET["id"])) {
     $id = $_GET["id"];
     $result = mysqli_query($conn, "DELETE FROM accessories WHERE id = '$id'");
